Refactored Spark UDFs
