+++ kuwala/core/neo4j/importer/requirements.txt
@@ -1,7 +1,11 @@
 arrow==1.1.0
+certifi==2021.5.30
+charset-normalizer==2.0.3
 dateparser==1.0.0
 findspark==1.4.2
+fuzzywuzzy==0.18.0
 h3==3.7.3
+idna==3.2
 moment==0.12.1
 neo4j==4.3.1
 numpy==1.20.3
@@ -11,10 +15,15 @@ py4j==0.10.9
 pycountry==20.7.3
 pyspark==3.1.2
 python-dateutil==2.8.1
+python-Levenshtein==0.12.2
 pytz==2021.1
 questionary==1.10.0
 regex==2021.4.4
+requests==2.26.0
+Shapely==1.7.1
 six==1.16.0
 times==0.7
 tzlocal==2.1
+urllib3==1.26.6
 wcwidth==0.2.5
+websocket-client==1.1.0
