+++ kuwala/pipelines/google-poi/requirements.txt
@@ -14,11 +14,7 @@ h3==3.7.2
 hpack==4.0.0
 Hypercorn==0.11.2
 hyperframe==6.0.1
-
-# Pip was unable to solve dependencies, suggested was to remove the version and let pip figure out the best version.
-idna
-#
-
+idna==2.10
 itsdangerous==1.1.0
 Jinja2==2.11.3
 MarkupSafe==1.1.1
@@ -38,6 +34,7 @@ pytz==2021.1
 Quart==0.14.1
 regex==2021.4.4
 requests==2.25.1
+Shapely==1.7.1
 six==1.16.0
 times==0.7
 toml==0.10.2
