+++ kuwala/pipelines/google-poi/src/app.py
@@ -2,7 +2,7 @@ import logging
 import os
 import sys
 
-sys.path.insert(0, '../../common/')
+sys.path.insert(0, '../../../common/')
 sys.path.insert(0, '../')
 
 from dotenv import load_dotenv
