+++ kuwala/pipelines/google-poi/src/pipeline/main.py
@@ -1,3 +1,8 @@
+import sys
+
+sys.path.insert(0, '../../../../common/')
+sys.path.insert(0, '../../')
+
 from SearchScraper import SearchScraper
 from search_string_generator import generate_search_strings
