+++ kuwala/pipelines/osm-poi/src/main.py
@@ -1,4 +1,9 @@
 import questionary
+import sys
+
+sys.path.insert(0, '../../../common/')
+sys.path.insert(0, '../')
+
 from Downloader import Downloader
 from Processor import Processor
