+++ kuwala/pipelines/population-density/dockerfile
@@ -1,9 +1,10 @@
 FROM rappdw/docker-java-python
 
-COPY ./pipelines/population-density /opt/app
+COPY ./common/python_utils /opt/common/python_utils
+COPY ./pipelines/population-density /opt/pipelines/app
 
-WORKDIR /opt/app
+WORKDIR /opt/pipelines/app
 RUN pip install -r requirements.txt
 
-WORKDIR /opt/app/src
+WORKDIR /opt/pipelines/app/src
 ENTRYPOINT [ "python", "main.py" ]
