+++ kuwala/pipelines/population-density/requirements.txt
@@ -11,12 +11,15 @@ charset-normalizer==2.0.3
 ckanapi==4.6
 click==8.0.1
 colorlog==5.0.1
+coverage==5.5
 cryptography==3.4.7
+cssselect==1.1.0
 decorator==5.0.9
 dnspython==2.1.0
 docopt==0.6.2
 email-validator==1.1.3
 et-xmlfile==1.1.0
+fuzzywuzzy==0.18.0
 greenlet==1.1.0
 h3==3.7.3
 hdx-python-api==5.1.8
@@ -26,41 +29,60 @@ html5lib==1.1
 idna==3.2
 ijson==3.1.4
 inflect==5.3.0
+iniconfig==1.1.1
 itsdangerous==2.0.1
 jmespath==0.10.0
 jsonlines==2.0.0
 jsonpath-ng==1.5.3
 libhxl==4.24.1
 linear-tsv==1.1.0
+lxml==4.6.3
 ndg-httpsclient==0.5.1
 num2words==0.5.10
+numpy==1.21.1
 openpyxl==3.0.7
+packaging==21.0
+pluggy==0.13.1
 ply==3.11
+pprintpp==0.4.0
 prompt-toolkit==3.0.19
+py==1.10.0
 py4j==0.10.9
 pyasn1==0.4.8
+pycountry==20.7.3
+pycountry-convert==0.7.2
 pycparser==2.20
 pyOpenSSL==20.0.1
+pyparsing==2.4.7
 pyphonetics==0.5.3
+pyquery==1.4.3
 pyspark==3.1.2
+pytest==6.2.4
+pytest-cov==2.12.1
+pytest-mock==3.6.1
 python-dateutil==2.8.2
 python-io-wrapper==0.2
+python-Levenshtein==0.12.2
 python-slugify==5.0.2
+pytz==2021.1
 PyYAML==5.4.1
 quantulum3==0.7.7
 questionary==1.10.0
 ratelimit==2.2.1
+repoze.lru==0.7
 requests==2.26.0
 requests-cache==0.7.1
 requests-file==1.5.1
 ruamel.yaml==0.17.10
 ruamel.yaml.clib==0.2.6
 s3transfer==0.5.0
+Shapely==1.7.1
 six==1.16.0
 soupsieve==2.2.1
 SQLAlchemy==1.4.21
 tabulator==1.53.5
 text-unidecode==1.3
+toml==0.10.2
 unicodecsv==0.14.1
 Unidecode==1.2.0
 url-normalize==1.4.3
