+++ kuwala/pipelines/population-density/src/Processor.py
@@ -1,21 +1,14 @@
 import math
-import h3
 import os
 import time
 from functools import reduce
 from pyspark.sql import SparkSession
-from pyspark.sql.functions import col, sum, udf
-from pyspark.sql.types import DoubleType, StringType
-
-DEFAULT_RESOLUTION = 11
+from pyspark.sql.functions import col, sum
+from pyspark.sql.types import DoubleType
+from python_utils.src.spark_udfs import get_h3_index
 
 
 class Processor:
-    @staticmethod
-    @udf(returnType=StringType())
-    def get_h3_index(lat: str, lng: str):
-        return h3.geo_to_h3(float(lat), float(lng), DEFAULT_RESOLUTION)
-
     @staticmethod
     def start(files: [dict], output_dir: str):
         memory = os.getenv('SPARK_MEMORY') or '16g'
@@ -44,7 +37,7 @@ class Processor:
             df = df \
                 .withColumnRenamed(df.columns[2], t) \
                 .withColumn(t, col(t).cast(DoubleType())) \
-                .withColumn('h3Index', Processor.get_h3_index(col(lat_column), col(lng_column))) \
+                .withColumn('h3Index', get_h3_index(col(lat_column), col(lng_column))) \
                 .drop(lat_column, lng_column) \
                 .groupBy('h3Index') \
                 .agg(sum(t).alias(t))
