+++ kuwala/pipelines/population-density/src/main.py
@@ -1,3 +1,8 @@
+import sys
+
+sys.path.insert(0, '../../../common/')
+sys.path.insert(0, '../')
+
 from Downloader import Downloader
 from Processor import Processor
